# public/conceptionRessources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "home-iot-api",
    "description": "The API for learning formation NodeJS project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/REMITAFFOREAU_1/LearningFormation/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/rental/findByRenter/{renterId}": {
      "get": {
        "tags": [
          "Rental"
        ],
        "summary": "Get rentals by renter",
        "operationId": "findRentalsByRenter",
        "parameters": [
          {
            "name": "renterId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_1"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
    "/rental/findByVehicle/{vehicleId}": {
      "get": {
        "tags": [
          "Rental"
        ],
        "summary": "Get rentals by vehicle",
        "operationId": "findRentalsByVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_1"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
    "/rental/getRentalDetails/{id}": {
      "get": {
        "tags": [
          "Rental"
        ],
        "summary": "Get rental detail",
        "operationId": "getRentalDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_3"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
    "/rental/update/{id}": {
      "put": {
        "tags": [
          "Rental"
        ],
        "summary": "Update rental status",
        "operationId": "updateRentalStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Rental"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Rental"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_4"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
...
    "/comment/{vehicleId}": {
      "get": {
        "tags": [
          "Comment"
        ],
        "summary": "Get comments for a vehicle",
        "operationId": "findCommentsByVehicle",
        "parameters": [
          {
            "name": "vehicleId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          },
          "404": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_404_5"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_400"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "r@yahoo.fr"
          },
          "pseudo": {
            "type": "string",
            "example": "coco"
          },
          "password": {
            "type": "string",
            "format": "hash",
            "example": "0729a580-2240-11e6-9eb5-0002a5d5c51b"
          },
          "role": {
            "type": "string",
            "example": "prestataire"
          }
        }
      },
      "CategoryVehicle": {
        "type": "object",
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Automobile"
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "libelle": {
            "type": "string",
            "example": "Renault Mégane"
          },
          "description": {
            "type": "string",
            "example": "Berling rouge de grand standing"
          },
          "imageUrl": {
            "type": "string",
            "example": "http://1230.fr"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryVehicle"
          },
          "priceDay": {
            "type": "integer",
            "format": "int32",
            "example": 12
          },
          "prestataire": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Rental": {
        "type": "object",
        "properties": {
          "vehicle": {
            "$ref": "#/components/schemas/Vehicle"
          },
          "renter": {
            "$ref": "#/components/schemas/User"
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2024-03-12"
          },
          "endDate": {
            "type": "string",
            "format": "date",
            "example": "2024-03-12"
          },
          "status": {
            "type": "string",
            "example": "pending"
          }
        }
      },
      "Comment": {
        "type": "object",
        "properties": {
          "vehicle": {
            "$ref": "#/components/schemas/Vehicle"
          },
          "author": {
            "$ref": "#/components/schemas/User"
          },
          "comment": {
            "type": "string",
            "example": "Très bonne voiture !"
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2024-03-12"
          }
        }
      },
...
      "inline_response_400": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "inline_response_404_1": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "No vehicle in this category."
          }
        }
      },
...
      "inline_response_404_3": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Vehicle not found."
          }
        }
      },
      "inline_response_404_4": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Rental not found."
          }
        }
      },
...
      "inline_response_404_5": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "No Comment for this vehicle."
          }
        }
      },
...
    }
  }
}